fix: throw user error when converted amount is too small
Previously, when the resulting converted amount of a conversion request
is too small (<=0n) we would throw a normal `Error`. However, this
results in a generic `Internal Server Error` message which isn't really
helpful. Since the user sent an amount which is too small to be
converted properly, we can return a `KeetaAnchorUserError` here which
will be converted into a "Bad Request" by the HTTP server SDK. This
clearly signals the user user that they made a mistake and should adjust
their requested amount.